feat(link): implement advanced in-app browser options
